# Makes liveness test faster