Makes the _"should \*not\* restart"_ liveness e2e test take 45 seconds instead of 4 minutes.

I measured the timing of this section of the e2e test, and the positive tests ("**should** restart") consistently take exactly 26 seconds. (See Jenkins history for this regularity.) Thus, 45 seconds is a good buffer, and drastically speeds up the test that always waits for the timeout.

Other minor improvements:
- print elapsed time during test
- change printing from `"pod %s in namespace %s", pod, ns` to the more concise `"pod %s/%s", ns, pod`
- increase polling interval (5s &rarr; 2s)
- use more precise types for printing (`%t` instead of `%v` for bools)