@mbforbes 2015-07-18
It appears 45s, while begin plenty long for normal runs, isn't long enough for parallel runs :-( Bumped to 1m.