@mbforbes 2015-07-18
This is good now. We've got two Jenkins flakes/failures:

```
[Fail] Services [It] should serve multiport endpoints from pods 
/go/src/github.com/GoogleCloudPlatform/kubernetes/_output/dockerized/go/src/github.com/GoogleCloudPlatform/kubernetes/test/e2e/service.go:959

[Fail] Services [It] should serve a basic endpoint from pods 
/go/src/github.com/GoogleCloudPlatform/kubernetes/_output/dockerized/go/src/github.com/GoogleCloudPlatform/kubernetes/test/e2e/service.go:959
```

The tests in question finish in either 28s or 10s.