@mbforbes 2015-07-18
@ixdy let me know if you've already got an issue for the flaking service test (I see flakes in the Jenkins `kubernetes-e2e-gce-parallel` job).