@mbforbes 2015-07-20
@k8s-bot test this please