@mbforbes 2015-07-21
@k8s-bot test this please