@ixdy 2015-07-23
ha, I was just starting to take a look at this!

I really wish there were a better way to do the negative test than wait for timeout.