@ixdy 2015-07-23
LGTM, thanks for the thorough fix, even the commented out tests!

(... wait, there are commented-out tests? augh!)